uppercase issue

+++ prismx/feature.py
@@ -49,12 +49,12 @@ def get_average_correlation(workdir: str, i: int, library: Dict, intersect: bool
     preds = []
     for ll in list(library.keys()):
         if intersect:
-            preds.append(correlation.loc[:, library[ll]].loc[ugenes,:].astype(np.float32).mean(axis=1))
+            preds.append(correlation.loc[:, library[ll]].loc[ugenes,:].mean(axis=1))
         else:
-            preds.append(correlation.loc[:, library[ll]].astype(np.float32).mean(axis=1))
+            preds.append(correlation.loc[:, library[ll]].mean(axis=1))
     features = pd.concat(preds, axis=1)
     features.columns = list(library.keys())
-    features = pd.DataFrame(features.fillna(0), dtype=np.float32)
+    features = pd.DataFrame(features.fillna(0), dtype=np.float16)
     features = features.reset_index()
     features.columns = features.columns.astype(str)
     features.to_feather(workdir+"/features/features_"+str(i)+".f")
